except search

+++ src/app/student-management/student-management.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { DialogService } from '../sharedFolder/dialog.service';
 import { Student } from '../sharedFolder/student.model';
 import { StudentService } from './student.service';
 
@@ -12,7 +11,7 @@ export class StudentManagementComponent implements OnInit {
   students: Student[] = [];
   title: string ='';
   headElements = ['ID', 'Name', 'College', 'Year', 'Department', 'email', 'Mobile', 'Action'];
-  constructor(private studentService: StudentService, private dialogService: DialogService) { }
+  constructor(private studentService: StudentService) { }
 
   ngOnInit(): void {
     this.studentService.studentChanged
@@ -24,21 +23,23 @@ export class StudentManagementComponent implements OnInit {
 
   onCreate(){
     this.title = 'Create';
-    this.dialogService.dialog.next(this.title);
+    this.studentService.createStudent.next(this.title);
   }
 
   onEdit(index: number) {
-    this.title = 'Edit';
-    this.dialogService.dialog.next(this.title);  
+    this.studentService.EditStudent.next(index);
   }
 
   onView(index: number) {
-    this.title = 'View';
-    this.dialogService.dialog.next(this.title);
+    this.studentService.studentSelected.next(index);
   }
 
   onDelete(index: number) {
-    this.studentService.deleteStudent(index)
+    if(confirm("Are you sure to delete?"))
+    this.studentService.deleteStudent(index);
   }
 
+  onSearch(table: any){
+    console.log(table);
+  }
 }
